requirements.txt
flask
sqlalchemy
psycopg2-binary
pdfplumber
python-docx
spacy
